Conversão de int para float

+++ Simplex.java
@@ -7,30 +7,30 @@ public class Simplex {
     public static void main(String[] args) throws IOException {
 
         String caminhoArquivo = "entrada.txt";
-        int[][] matriz = lerArquivoEConstruirMatriz(caminhoArquivo);
+        float[][] matriz = lerArquivoEConstruirMatriz(caminhoArquivo);
 
         imprimirMatriz(matriz);
 
-       gaussJordan(matriz);
+        gaussJordan(matriz);
     }
 
-    public static int[][] lerArquivoEConstruirMatriz(String caminho) throws IOException {
+    public static float[][] lerArquivoEConstruirMatriz(String caminho) throws IOException {
         BufferedReader br = new BufferedReader(new FileReader(caminho));
 
-        int tipo = Integer.parseInt(br.readLine().trim());
+        float tipo = Float.parseFloat(br.readLine().trim());
         int x = Integer.parseInt(br.readLine().trim());     
         int y = Integer.parseInt(br.readLine().trim());     
 
         String[] linhaFO = br.readLine().trim().split("\\s+");
-        int[] funcObjetivo = new int[x];
+        float[] funcObjetivo = new float[x];
 
         for (int i = 0; i < x; i++) {
-            funcObjetivo[i] = Integer.parseInt(linhaFO[i]) * tipo;
+            funcObjetivo[i] = Float.parseFloat(linhaFO[i]) * tipo;
         }
 
         int colunas = x + y + 1;
         int linhas = y + 1;
-        int[][] matriz = new int[linhas][colunas];
+        float[][] matriz = new float[linhas][colunas];
 
         for (int i = 0; i < x; i++) {
             matriz[0][i] = -funcObjetivo[i];
@@ -40,18 +40,18 @@ public class Simplex {
             String[] restr = br.readLine().trim().split("\\s+");
 
             for (int j = 0; j < x; j++) {
-                matriz[i][j] = Integer.parseInt(restr[j]);
+                matriz[i][j] = Float.parseFloat(restr[j]);
             }
 
-            matriz[i][x + (i - 1)] = 1;
-            matriz[i][colunas - 1] = Integer.parseInt(restr[x]);
+            matriz[i][x + (i - 1)] = 1f;
+            matriz[i][colunas - 1] = Float.parseFloat(restr[x]);
         }
 
         br.close();
         return matriz;
     }
 
-    public static void gaussJordan(int matriz[][]) {
+    public static void gaussJordan(float matriz[][]) {
         while (true) {
             imprimirMatriz(matriz);
 
@@ -64,7 +64,7 @@ public class Simplex {
                 return;
             }
 
-            int pivo = matriz[linhaPivo][colunaPivo];
+            float pivo = matriz[linhaPivo][colunaPivo];
 
             for (int j = 0; j < matriz[linhaPivo].length; j++) {
                 matriz[linhaPivo][j] /= pivo;
@@ -72,7 +72,7 @@ public class Simplex {
 
             for (int i = 0; i < matriz.length; i++) {
                 if (i != linhaPivo) {
-                    int fator = matriz[i][colunaPivo];
+                    float fator = matriz[i][colunaPivo];
                     for (int j = 0; j < matriz[i].length; j++) {
                         matriz[i][j] -= fator * matriz[linhaPivo][j];
                     }
@@ -86,7 +86,7 @@ public class Simplex {
         imprimirSolucao(matriz);
     }
 
-    public static void imprimirSolucao(int[][] matriz) {
+    public static void imprimirSolucao(float[][] matriz) {
         int linhas = matriz.length;
         int colunas = matriz[0].length;
 
@@ -99,10 +99,10 @@ public class Simplex {
             int linhaBase = -1;
 
             for (int i = 1; i < linhas; i++) {
-                if (matriz[i][j] == 1) {
+                if (matriz[i][j] == 1f) {
                     boolean ehBase = true;
                     for (int k = 1; k < linhas; k++) {
-                        if (k != i && matriz[k][j] != 0) {
+                        if (k != i && matriz[k][j] != 0f) {
                             ehBase = false;
                             break;
                         }
@@ -122,10 +122,10 @@ public class Simplex {
             int linhaBase = -1;
 
             for (int i = 1; i < linhas; i++) {
-                if (matriz[i][j] == 1) {
+                if (matriz[i][j] == 1f) {
                     boolean ehBase = true;
                     for (int k = 1; k < linhas; k++) {
-                        if (k != i && matriz[k][j] != 0) {
+                        if (k != i && matriz[k][j] != 0f) {
                             ehBase = false;
                             break;
                         }
@@ -143,13 +143,12 @@ public class Simplex {
             }
         }
 
-        // Valor da função objetivo
         System.out.println("\nValor ótimo = " + matriz[0][colunas - 1]);
     }
 
-    public static void imprimirMatriz(int matriz[][]) {
-        for (int[] linha : matriz) {
-            for (int v : linha) {
+    public static void imprimirMatriz(float matriz[][]) {
+        for (float[] linha : matriz) {
+            for (float v : linha) {
                 System.out.print(v + "\t");
             }
             System.out.println();
@@ -157,9 +156,9 @@ public class Simplex {
         System.out.println();
     }
 
-    public static int obterColunaPivo(int matriz[][]) {
+    public static int obterColunaPivo(float matriz[][]) {
         int colunaPivo = 0;
-        int valorMinimo = matriz[0][0];
+        float valorMinimo = matriz[0][0];
 
         for (int j = 1; j < matriz[0].length - 1; j++) {
             if (matriz[0][j] < valorMinimo) {
@@ -170,13 +169,13 @@ public class Simplex {
         return colunaPivo;
     }
 
-    public static int obterLinhaPivo(int matriz[][], int colunaPivo) {
+    public static int obterLinhaPivo(float matriz[][], int colunaPivo) {
         int linhaPivo = -1;
-        double razaoMinima = Double.MAX_VALUE;
+        float razaoMinima = Float.MAX_VALUE;
 
         for (int i = 1; i < matriz.length; i++) {
             if (matriz[i][colunaPivo] > 0) {
-                double razao = (double) matriz[i][matriz[i].length - 1] / matriz[i][colunaPivo];
+                float razao = matriz[i][matriz[i].length - 1] / matriz[i][colunaPivo];
                 if (razao < razaoMinima) {
                     razaoMinima = razao;
                     linhaPivo = i;
